wobble
Added wobble effect. The wobble effect makes the branches (line segments) jagged. Each branch is separated into 10 line segments. The angle of each segment is different. The angle usually changes linearly, but the wobble effect adds a random number onto that. But these random number can add up, the the wobble effect was causing the branches to turn around below the trunk. So I added a counter wobble. For every wobble, I added an equal and opposite wobble to the segment after it, cancelling out the wobbles so that the angle ended up much closer to the target angle than without the counter wobble.