tierSmoothness and segmentExtend
I smoothed the thickness of the branches between the tiers. To do that, I had to switch to drawing each segment as a separate line. But that left gaps between the lines. To fix that, I added another property; segment extend. It makes the 10 segments that make up the branches longer, overextending them.