spacing and scaling
I made the tree spacing and scaling better. The x and y are in terms of percentage of a rectangle, which used to be the window. Now, it's the biggest square that fits inside the window, and its side-lengths are represented as "m", meaning "multiplier". This prevents the trees from being stretched based on the width and height of the canvas. I also made the spacing of the trees better by creating a "tree" class and organizing the trees by y-position in the "trees" array, so that the farther away trees are drawn first so the closer trees are layered in front. I also made "mp", which means "multiplier proportion". The tree size is proportional to the m-square, but the trees are spread out over the canvas, which is larger than the m-square. So mp is the proportion of the m-square to the canvas. The amount of trees is multiplied by mp so that there will always be the same average tree density.