branch and segment length vary
Added branchSegments, the number of segments per branch, so that it can be numbers other than 10. Added branchSegmentsVary, which varies the number of segments in a particular branch. Added "s" which means "segments" and it represents the number of segments in a particular branch. Added branchSegmentLengthVary, which varies the length of a particular segment. Added branchLengthVary, which varies the length of a particular branch. The length of a particular branch is represented by "bm", meaning "branch multiplier". I also made the trees in front brighter.