fixed leaf points
I used ctx.lineTo() for every point on the leaves so it filled in the gaps between the leaf's points. To fit that, I switched to drawing two triangles per leaf point. I also made ld.leafClumpSpread work and modified it so that its more likely for the leaves to be closer to the top of the tree.